+++ boot/malloc.h
@@ -3,7 +3,7 @@
 #include "types.h"
 
 void *malloc(uint16_t bytes);
-void *calloc(uint16_t num, uint16_t size);
+void *calloc(unsigned num, unsigned size);
 void free(void *p);
 
 uint16_t far_malloc(uint32_t bytes);
